Added emotion recognition application
This commit adds all required files to the application for emotion recognition.

The training procedure is the responsibility of the webcamcnn.py script. It verifies GPU availability, which can accelerate the training process if CUDA-enabled hardware is present. The script imports a dataset of grayscale images classified by seven emotions, preprocesses the images, and defines the structure of a CNN model. The model is then trained with an Adam optimizer and a cross-entropy loss function that is categorical. Mechanisms for early halting and model checkpointing are in place to prevent overfitting and preserve the optimal model. Additionally, data augmentation techniques are utilized to improve model generalization.

Using the webcamtest.py script, the emotion recognition model is evaluated in real-time. It captures video frames from a webcam, preprocesses the captured images, and then transmits them to the trained model for emotion prediction. The predicted emotion is then displayed on the console, enabling emotion recognition in real-time.

The webcamcnn.h5 file contains the trained model generated by the webcamcnn.py script. The model file is stored in HDF5 format, allowing it to be easily loaded and utilized for prediction in a variety of environments. This pre-trained model enables users to execute the testing script without the computationally intensive need to train the model from inception.